==> 02-modulo02_sprint_01/02-pandas/02-pandas_2.ipynb <==
### Pair Programming Pandas II

# Lo primero, importamos las librerías:

import pandas as pd
import pickle
import numpy as np


1. Id cargando todos los ficheros en vuestro jupyter.

Para poner el práctica el read_clipboard elegir alguna tabla que os podaís encontrar en google y usad dicho método para traer los datos al jupyter notebook.


#pickle
archivo_pickle=  pd.read_pickle('datos/02-algas.pickle')
archivo_pickle

- .csv

df_csv = pd.read_csv("datos/quejas.csv", encoding="latin", sep = ";" )
df_csv.head()

- .json

df_json = pd.read_json("datos/quejas.json")
df_json.head()

- .xlsx (Excel)

df_xlsx = pd.read_excel("datos/quejas2.xlsx")
df_xlsx.head()

Para poner el práctica el read_clipboard elegid alguna tabla que os podaís encontrar en google y usad dicho método  para traer los datos al jupyter notebook.

archivo_clipboard = pd.read_clipboard()
archivo_clipboard

Una vez que los hayáis cargado, guardad esos datos en los siguientes formatos:

- csv
- excel
- json
- picke

df_csv.to_csv("datos/tabla.csv")

df_xlsx.to_excel("datos/tabla.xlsx")


df_json.to_json("datos/tabla.json")

archivo_pickle.to_pickle("datos/tabla.pickle")
