==> 02-modulo02_sprint_01/02-pandas/03-pandas_3.ipynb <==
### Pair Programming Pandas III

# Lo primero, importamos la librería:

import pandas as pd

1. Cargar el dataset
💡 Pista. Si hacemos la carga del dataset sin el parámetro encoding nos puede dar error (puede que no os de error, si es así menos problemas!).  tenéis una lista de los encodings disponibles en pandas. Buscad por los encoding de tipo ISO.



# Lo conseguimos con pd.read:

df_csv = pd.read_csv('datos/03-attacks.csv', sep = ",")
df_csv.head(10)


2. Haced una exploración básica del DataFrame:
- ¿Cuántas filas y columnas tiene el DataFrame?

df_csv.shape

print("Tiene 6114 filas y 22 collumnas.")

- Mostrad las primeras cinco filas y las últimas cinco filas del DataFrame.

# Las primeras cinco filas con head.:

df_csv.head(5)

# Las últimas cinco filas con .tail:

df_csv.tail(5)

- Mostrad 2 filas al azar del DataFrame.

df_csv.sample (2)

- ¿Cuál es el nombre de las columnas del DataFrame?

df_csv.columns

- ¿Cuáles son los principales estádisticos del DataFrame? Mostrad por separado las variables categóricas y las variables numéricas.


# Lo averiguamos usando .dtypes:

df_csv.dtypes 

# Las variables numéricas usando .describe():

df_csv.describe()

# Las variables categóricas pidiendo que incluya "objects":

df_csv.describe(include = 'object')

- ¿Cuál es el porcentaje de valores nulos en cada columna?

porcentaje_nulos = ((df_csv.isnull().sum() / len(df_csv))*100)
print(porcentaje_nulos)

- Cambiad el nombre de las columnas para que todas estén en minúsculas y reemplazad los espacios por _.


df_csv_mayus = df_csv.rename(columns = {'pdf':'PDF','href formula':'HREF Formula', 'href':'HREF','original order':'Original Order'})
df_csv_mayus

- Eliminad las columnas de Unnamed: 22 y Unnamed: 23, Case Number.1, Case Number.2, original order, Investigator or Source, pdf, href formula.


# Primera forma de hacerlo, llamando al DF y usando un drop sobre los nombres de las columnas:

df_borrada = df_csv_mayus.drop(['Unnamed: 22','Unnamed: 23','Case Number.1','Case Number.2','Original Order','Investigator or Source','PDF','HREF Formula'], axis = 1 )


df_borrada.head(1)

# Segunda forma de hacerlo, sacar los nombres de las columnas con .columns y sustituirlas en el DF:

df_csv_mayus.columns

df_columnas_new = df_csv_mayus.columns

df_columnas_new.drop(['Unnamed: 22','Unnamed: 23','Case Number.1','Case Number.2','Original Order','Investigator or Source','PDF','HREF Formula'])

- Establece la columna Case Number como índice del DataFrame.

df_borrada.set_index(["Case Number"], inplace = True)

df_borrada.index

df_borrada.head(1)

- Guarda el csv que no tiene esas columnas para seguir usándolo mañana.

df_borrada.to_csv('datos/03-attacks_new1.csv')
