==> 02-modulo02_sprint_01/02-pandas/04-pandas_4.ipynb <==
### Pair Programming Pandas IV

# HLo primero, importamos las librerías:

import pandas as pd
import numpy as np

1. Cargad el csv que guardasteis ayer.

df_tiburones = pd.read_csv("datos/03-attacks_new1.csv")
df_tiburones.head(3)

Antes de seguir con las preguntas, este DataFrame está terriblemente sucio. En las próximas sesiones lo iremos limpiando. Para evitar posibles errores con los nulos vamos a eliminar todos los nulos del DataFrame.

Para ello vamos a utilizar el método .dropna() de pandas. Tendréis que ejecutar la siguiente línea de código para eliminar los nulos del DataFrame y poder seguir haciendo el resto de los ejercicios.

df_tiburones.info()

df_total = df_tiburones.shape
df_total

- ¿Cuántas filas se han eliminado con esta línea de código?

df_tiburones.dropna(inplace = True)
df_tiburones

df_tiburones.shape

print("Se han borrado", df_total[0]-df_tiburones.shape[0], "filas en la eliminación")

2. ¿Cuál es el número de ataques que se han producido en Estados Unidos?

df_ataque_usa = df_tiburones[df_tiburones["Country"] == "USA"]
df_ataque_usa.head(3)

df_ataque_usa.shape

print(f'En USA se han producido {(df_ataque_usa.shape[0])} ataques')

3. ¿Y en España?

df_spain = df_tiburones[df_tiburones["Country"] == "SPAIN"]
df_spain.head(2)

df_spain.shape

print(f'En España se han producido {(df_spain.shape[0])} ataques')

- ¿En qué regiones se han producido los ataques (columna location)?

df_spain_location = df_spain["Location"].unique()
df_spain_location

- ¿Cuántos ataques se han producido en cada área (columna area)?

df_spain_ataques = df_spain["Area"].value_counts()
df_spain_ataques

- Extraed un DataFrame nuevo con los ataques que se han producido en Australia y Estados Unidos.

    📌 Nota Este ejercicio se puede resolver usando dos condiciones de filtrado o usando el método isin(). Hacedlo con ambas herramientas.


df_aus_usa = df_tiburones[(df_tiburones["Country"] == "USA") | (df_tiburones["Country"] == "AUSTRALIA")]
df_aus_usa.head(2)

df_aus_usa_isin = df_tiburones[(df_tiburones["Country"] == "USA").isin(df_tiburones["Country"] == "AUSTRALIA")]
df_aus_usa_isin.head(2)

- ¿Cuántos ataques se han producido en cada país?


df_usa_count = df_aus_usa["Country"].value_counts()
df_usa_count

- ¿Ha habido ataques en Australia a gente que estuviera remando (lo encontrareis como Paddling en el DataFrame en la columna activity)?

df_aus_unique = df_aus_usa["Country"].unique()
df_aus_unique

df_aus_paddling = df_aus_usa[(df_aus_usa["Country"] == "AUSTRALIA") & (df_aus_usa["Activity"] == "Paddling")]
df_aus_paddling.head()

df_aus_paddling.value_counts()

A la vista de los resultado vemos que no ha habido casos

- ¿Cuántos ataques se han producido en Australia antes del 2000?

ataque_antes_2000 = df_aus_usa[(df_aus_usa["Country"] == "AUSTRALIA") & (df_aus_usa["Year"] < 2000)]
print(f'Antes del 2000 hubo {ataque_antes_2000.shape[0]} ataques.')

- ¿Y después del 2000?


ataque_despues_2000 = df_aus_usa[(df_aus_usa["Country"] == "AUSTRALIA") & (df_aus_usa["Year"] >= 2000)]
print(f'Despues del 2000 hubo {ataque_despues_2000.shape[0]} ataques.')

ataque_despues_2000 = df_tiburones[(df_tiburones["Sex "] == "F")]

5. Extraed unicamente los ataques a mujeres.

df_ataques = df_tiburones[df_tiburones["Sex "] == "F"]
df_ataques.head(2)

df_tiburones.columns

6. ¿Cuántas mujeres fueron atacadas en España?


df_ataques_españa = df_ataques[(df_ataques["Country"] == "SPAIN") & (df_ataques["Sex "] == "F")]
df_ataques_españa

7. Extraed los ataques que se hayan producido en el año 2000, 1997 y 2006.

    Utilizad el método isin() para resolver este ejercicio.

df_ataques_fechas_mix = df_tiburones[df_tiburones["Year"].isin([1997, 2000, 2006])]
df_ataques_fechas_mix.head(2)

8. Si extraemos los valores únicos de la columna species nos podemos dar cuenta que está un poco sucia y extraer informacion útil de la misma se puede convertir en algo difícil.

    En este ejercicio tendréis que extraer los ataques que han ocurrido por tiburón blanco (lo encontraremos en la columna species como algo que contenga white o White).

    Usad el método str.contains() para resolver este ejercicio.

df_tiburonblanco = df_tiburones[df_tiburones["Species "].str.contains("White", "white", regex=True)]
df_tiburonblanco.head(2)

9. Extraed ahora los ataques por tiburón blanco o tigre.

    Utilizad el método str.contains() para resolver este ejercicio.

df_tiburonblanco_tigre = df_tiburones[df_tiburones["Species "].str.contains("White|white|Tiger|tiger", regex=True)]
df_tiburonblanco_tigre.head(2)

10. Utilizando la columna de year cread una nueva columna llamada siglo que tiene siglo-XX o siglo-XXI. Donde el año sea menor que 2000 pondremos siglo-XX y el resto de los años siglo-XXI.


df_tiburones["Siglo"] = np.where(df_tiburones["Year"] <= 2000, "siglo-XX", "siglo-XXI")

df_tiburones["Siglo"].unique()

11. Guardad como un archivo csv nuevo. 

df_tiburones.to_csv("datos/04-attacks.csv")
