==> 02-modulo02_sprint_01/02-pandas/05-pandas_5.ipynb <==
### Pair Programming Pandas V  

# Lo primero, importamos las librerías:

import pandas as pd
import numpy as np
import os
from os import getcwd

1. Cargad los tres ficheros nombrados anteriormente y los exploramos para familizarizarnos con ellos y saber que información tienen en común y debatir con tu compañera como podríamos juntar toda la información.

df_t = pd.read_csv("datos/04-attacks.csv",index_col=0)
df_t.head(3)

df_t_h = pd.read_csv("datos/05-attacks_historico.csv", index_col = 0)
df_t_h.head(3)

df_t_a = pd.read_csv("datos//05-attacks_adicional.csv", index_col = 0)
df_t_a.head(3)

Chequead que los nombres de las columnas ver si se llaman iguales entre dataframes.

### Describe

nombre_tibu3_des = df_t.describe().T
nombre_tibuhistorico_des = df_t_h.describe().T
nombre_tibuadicional_des = df_t_a.describe().T

nombre_tibu3_des

nombre_tibuhistorico_des

nombre_tibuadicional_des

### Columns

nombre_tibu3_col = df_t.columns
nombre_tibuhistorico_col = df_t_h.columns
nombre_tibuadicional_col = df_t_a.columns

nombre_tibu3_col

nombre_tibuhistorico_col

nombre_tibuadicional_col

### Shape

nombre_tibu3_shape = df_t.shape
nombre_tibuhistorico_shape = df_t_h.shape
nombre_tibuadicional_shape = df_t_a.shape

nombre_tibu3_shape

nombre_tibuhistorico_shape

nombre_tibuadicional_shape

### Borramos las columnas Unnamed

df_t = df_t.drop(["Unnamed: 0"], axis = 1)
df_t.head(1)

df_t_h = df_t_h.drop(["Unnamed: 22", "Unnamed: 23"], axis = 1)
df_t_h.head(1)

### Cambiamos el nombre a las columnas

lista_columnas = {col: col.strip().replace(" ", "_").lower() for col in df_t.columns}
df_t.rename(columns=lista_columnas, inplace=True)

df_t.head(1)


lista_columnas_historico = {col : col.strip().replace(" ", "_").lower() for col in df_t_h.columns}
df_t_h.rename(columns=lista_columnas_historico, inplace=True)

df_t_h.head(1)

lista_columnas_adicional = {col: col.strip().replace(" ", "_").lower() for col in df_t_a.columns}
df_t_a.rename(columns=lista_columnas_adicional, inplace=True)

df_t_a.head(1)

- ¿Están todas en minúsculas? ¿Tienen espacios?

    Están corregidas y bien, como podemos ver:

df_t.columns

df_t_h.columns

df_t_a.columns

En este momento nos habremos dado cuenta que el DataFrame que generamos 
en el pair programming Pandas IV es similar al DataFrame de los datos historicos.
Si recordamos, eliminamos algunas columnas en la sesión III que no eran útiles,
pero esas columnas están todavía en el histórico. Si no fuera por eso podríamos 
unir esos dataframes. Realizad los cambios que creais necesarios para que los
dataframes sean iguales y luego juntadlos usando el mejor método.

# Comprobamos cómo son las columnas y decidimos hacer primero un pd.concat y luego un merge:

df_t.columns

df_t_h.columns

df_t_a.columns

df_concat = pd.concat([df_t,df_t_h], axis=0)
df_concat.head(2)

Es el momento de unir la información de todos los ataques de tiburón que acabamos de crear con el DataFrame de información adicional. Debatid que método es el más correcto para unir toda esta información y ejecutadlo.

df_fin = df_concat.merge(df_t_a, how = "left", on= "case_number")
df_fin.head(2)

df_fin.columns

df_fin.shape

df_fin.describe()

df_fin.to_csv("datos/05-attacks_union.csv")

leer = pd.read_csv("datos/05-attacks_union.csv", index_col=0)
leer.head(2)
