==> 02-modulo02_sprint_01/02-pandas/07-pandas_7.ipynb <==
# importamos todas las librerías que necesitamos

import pandas as pd
pd.options.display.max_columns = None
import re
import regex as re
import numpy as np
import datetime

### Pair Programming Pandas VII

En el ejercicio de ayer estuvimos viendo el groupby y pudimos observar que los datos estaban un poco sucios... 😭. En los ejercicios de hoy pondremos en práctica los apply y nos crearemos algunas funciones que nos van a permitir limpiar un poco los datos, para que podamos extraer conclusiones de una forma más sencilla.

df = pd.read_csv("datos/05-attacks_union.csv", index_col = 0)
df.head(5)

### 1. Cread una columna nueva y una función que nos dé el mes en el que ocurrió el ataque. 

Tened en cuenta que no todas las filas tienen la misma estructura 
y que puede que no haya la información de mes. En ese caso devolved un nulo (NaN).
💡 Pista 💡Podéis usar regex.

date_unique = df["date"].unique()
date_unique

def fecha(x):
    try:
        return re.findall('([A-Z]{1}[a-z]{2})\-',x)[0]
    except:
        return np.nan

df['mes'] = df['date'].apply(fecha)

df.head(2)

### 2. Vamos con la columna fatal_(y/n). 
Lo primero que tenéis que hacer es evaluar los valores únicos que hay. Esperaríamos tener solo dos valores: y y n, pero tristemente no será así 😭. Cread una columna nueva y una función que devuelva únicamente y o n y que devuelva un nulo (NaN) si no se encuentra el valor.


df["fatal_(y/n)"].unique()

def yn(x):
    if x == "N":
        return "N"
    elif x == "Y":
        return "Y"
    elif x == " N":
        return "N"
    elif x == "N ":
        return "N"
    elif x == "y":
        return "Y"
    else:
        return np.nan

df["fatal"] = df["fatal_(y/n)"].apply(yn)
df.head(2)

df["fatal"].unique()

### 3. Seguid la misma lógica que en el ejercicio anterior pero con la columna sex. 
Pero en este caso vamos a incluir un retito más: En este caso querremos tener dos valores únicos F y M. De nuevo en la columna hay otros carácteres raros como 'M ', 'lli', 'N', '.'. En esos casos, queremos que se reemplacen por M o F de forma aleatoria (tendréis que usar la librería random). En caso de que sea nulo, que la función devuelva nulo (np.nan).


df["sex"].unique()

def reemplazar_mf(x):
    if x == "F":
        return "F"
    elif x == "M":
        return "M"
    else:
        return np.nan

df["sex_fm"] = df["sex"].apply(reemplazar_mf)
df.head(5)

df["sex_fm"].unique()

### 4. Una vez que hayáis terminado eliminad las columnas originales conservando las nuevas.


df.drop(["fatal_(y/n)", "sex"], axis = 1, inplace = True)

df.head(5)

### 5. Guardad el dataframe en un csv que usaremos más adelante.

df.to_csv("datos/07-attack.csv")
